+++ odoo/local-src/specific_stock/__manifest__.py
@@ -14,6 +14,7 @@
         'product_expiry',
         'purchase',
         'sale',
+        'sale_stock',
         'specific_data',
         'specific_helpdesk',
         'specific_purchase',